feat(view): add the feature of reading size of bridge
